[tests] Use test-runner to run unhandled-exception tests
